requirements.txt
faiss-cpu
groq
langchain-groq
PyPDF2
langchain_google_genai
langchain
streamlit
langchain_community
python-dotenv
pypdf
google-cloud-aiplatform>=1.38
spacy
beautifulsoup4
selenium
webdriver-manager
requests